+++ plugins/at.ac.tuwien.big.momot.core/src/at/ac/tuwien/big/momot/reactive/IReactiveSearch.java
@@ -13,6 +13,6 @@ public interface IReactiveSearch {
    public SearchResult performSearch(final EGraph graph, final String algorithmName, String experimentName, int run,
          final int evaluations, final TerminationCondition terminationCondition, final int solutionLength,
          final int populationSize, final List<ITransformationVariable> reinitSeed, final double reinistBestObj,
-         float reinitPortion, boolean recordBestObjective);
+         double reinitPortion, boolean recordBestObjective);
 
 }
